chnages on onboarding test scripts
