+++ cypress-tests/cypress/constants/selectors/common.js
@@ -9,6 +9,7 @@ export const commonSelectors = {
     '.drawer-container > [style="position: fixed; z-index: 9999; inset: 16px; pointer-events: none;"] > .go4109123758 > .go2072408551',
   toastCloseButton: '[data-cy="toast-close-button"]',
   editButton: "[data-cy=edit-button]",
+  fileSelector: "[data-cy=uploaded-file-data]",
   searchField: "[data-cy='widget-search-box-search-bar']",
   firstWidget: "[data-cy=widget-list]:eq(0)",
   canvas: "[data-cy=real-canvas]",
@@ -65,7 +66,7 @@ export const commonSelectors = {
   signInHeader: '[data-cy="sign-in-header"]',
   signInSubHeader: '[data-cy="signup-info"]',
   createAnAccountLink: '[data-cy="create-an-account-link"]',
-  SignUpSectionHeader: '[data-cy="sign-up-header"]',
+  signUpSectionHeader: '[data-cy="sign-up-header"]',
   signInRedirectText: '[data-cy="signin-redirect-text"]',
   signInRedirectLink: '[data-cy="signin-link"]',
   signUpTermsHelperText: '[data-cy="signup-terms-helper"]',
@@ -75,8 +76,9 @@ export const commonSelectors = {
   invitePageHeader: '[data-cy="join-my-workspace-header"]',
   invitePageSubHeader: '[data-cy="onboarding-page-description"]',
   userNameInputLabel: '[data-cy="name-input-label"]',
-  invitedUserName: '[data-cy="name-input-value"]',
+  invitedUserName: '[data-cy="name-input-input-value"]',
   invitedUserEmail: '[data-cy="email-input-value"]',
+  invitedUseremail: '[data-cy="email-input-input-value"]',
   acceptInviteButton: '[data-cy="accept-invite-button"]',
   databaseIcon: '[data-cy="icon-database"]',
   profileSettings: '[data-cy="profile-settings"]',
@@ -109,8 +111,8 @@ export const commonSelectors = {
   avatarImage: '[data-cy="avatar-image"]',
   selfHostSetUpBanner: '[data-cy="setup-banner-inner"]',
   HostBanner: '[data-cy="onboarding-image"]',
-  AdminSetup: '[data-cy="set-up-your-admin-account-header"]',
-  SignupTerms: '[data-cy="signup-terms-helper"]',
+  adminSetup: '[data-cy="set-up-your-admin-account-header"]',
+  signupTerms: '[data-cy="signup-terms-helper"]',
   selfHostSetUpSubBanner: '[data-cy="setup-sub-banner"]',
   selfHostSetUpCard: '[data-cy="setup-card"]',
   selfHostSetUpCardImage: '[data-cy="setup-card-image"]',
@@ -139,9 +141,9 @@ export const commonSelectors = {
   companyNameInputField: '[data-cy="company-name-input-field"]',
   nameInputField: '[data-cy="name-input-input"]',
   emailInputField: '[data-cy="email-input-input"]',
-  Skipbutton: '[class="driver-close-btn"]',
-  BackLogo: '[data-cy="editor-page-logo"]',
-  Backtoapps: '[data-cy="back-to-app-option"]',
+  skipbutton: '[class="driver-close-btn"]',
+  backLogo: '[data-cy="editor-page-logo"]',
+  backtoapps: '[data-cy="back-to-app-option"]',
   signUpButton: '[data-cy="sign-up-button"]',
   emailImage: '[data-cy="email-image"]',
   spamMessage: '[data-cy="info-message"]',
