+++ cypress-tests/cypress/constants/selectors/onboarding.js
@@ -5,16 +5,16 @@ export const cyParamName = (paramName = "") => {
 export const onboardingSelectors = {
   emailLabel: "[data-cy='email-label']",
   emailInput: "[data-cy='email-input-input']",
-  SignupEmailInput: "[data-cy='email-input']",
-  Loginemailinput: "[data-cy='email-input']",
+  signupEmailInput: "[data-cy='email-input']",
+  loginEmailInput: "[data-cy='email-input']",
   passwordLabel: "[data-cy='password-label']",
   passwordInput: "[data-cy='password-input-input']",
-  LoginpasswordInput: "[data-cy='password-input']",
+  loginPasswordInput: "[data-cy='password-input']",
   signInButton: "[data-cy='sign-in-button']",
   passwordError: "[data-cy='password-error']",
   nameInput: "[data-cy='name-input']",
   nameLabel: '[data-cy="name-input-label"]',
-  SignupNameLabel: '[data-cy="name-label"]',
-  ForgotEmailInput: '[data-cy="email-input-field-input"]',
+  signupNameLabel: '[data-cy="name-label"]',
+  forgotEmailInput: '[data-cy="email-input-field-input"]',
   createAnAccountLink: '[data-cy="create-an-account-link"]',
 }
