+++ cypress-tests/cypress/constants/texts/common.js
@@ -59,7 +59,7 @@ export const commonText = {
   loginButton: " Login",
   signInHeader: "Sign in",
   signInSubHeader: "New to ToolJet?Create an account",
-  SignUpSectionHeader: "Join ToolJet",
+  signUpSectionHeader: "Join ToolJet",
   signInRedirectText: "Already have an account?",
   signInRedirectLink: "Sign in",
   signUpTermsHelperText: "By signing up you are agreeing to the",
@@ -95,7 +95,7 @@ export const commonText = {
   continueButton: "Continue",
   resetPasswordButton: "Reset password",
   setUpWorkspaceHeader: "Set up your workspace",
-  LoginpasswordLabel: "Password *",
+  loginPasswordLabel: "Password *",
   userRolePageHeader: "What best describes your role?",
   sizeOftheCompanyHeader: "What is the size of your company?",
   workspaceNameInputLabel: "Workspace name *",
