+++ cypress-tests/cypress/fixtures/bulkUser/10usersupload.csv
@@ -1,0 +1,11 @@
+First Name,Last Name,Email,User Role,Group
+test1,user,test1@gmail.com,Builder,
+test2,user,test3@gmail.com,End User,
+Test3,Example,test12@gmail.com,Admin,
+Test4,Example,test23@gmail.com,End User,
+Test5,Example,test24@gmail.com,End User,
+Test6,Example,test25@gmail.com,End User,
+Test7,Example,test26@gmail.com,Admin,
+Test8,Example,test27@gmail.com,Builder,
+Test9,Example,test28@gmail.com,Builder,
+Test10,Example,test29@gmail.com,End User,
