+++ cypress-tests/cypress/support/utils/manageUsers.js
@@ -158,7 +158,7 @@ export const inviteUser = (firstName, email) => {
 
 export const confirmInviteElements = (email) => {
 
-  cy.get(commonSelectors.SignUpSectionHeader).verifyVisibleElement(
+  cy.get(commonSelectors.signUpSectionHeader).verifyVisibleElement(
     "have.text", "Sign up");
   cy.get('[data-cy="signup-info"]').verifyVisibleElement(
     "have.text", "Sign up to the workspace - My workspace. ");
@@ -213,9 +213,9 @@ export const bulkUserUpload = (file, fileName, toastMessage) => {
   });
   cy.get(usersSelector.uploadedFileData).should("contain", fileName);
   cy.get(usersSelector.buttonUploadUsers).click();
-  cy.get(commonSelectors.newToastMessage)
-    .should("be.visible")
-    .and("have.text", toastMessage);
+  // cy.get(commonSelectors.newToastMessage)
+  //   .should("be.visible")
+  //   .and("have.text", toastMessage);
   cy.get(usersSelector.toastCloseButton).click();
 
   cy.wait(200);
@@ -245,7 +245,7 @@ export const copyInvitationLink = (firstName, email) => {
 export const fillUserInviteForm = (firstName, email) => {
   cy.get(usersSelector.buttonAddUsers).click();
   cy.clearAndType(onboardingSelectors.nameInput, firstName);
-  cy.clearAndType(onboardingSelectors.SignupEmailInput, email);
+  cy.clearAndType(onboardingSelectors.signupEmailInput, email);
 };
 
 export const selectUserGroup = (groupName) => {
@@ -296,7 +296,7 @@ export const inviteUserWithUserGroups = (
   cy.wait(1000);
   fetchAndVisitInviteLink(email);
   cy.wait(2000);
-  cy.clearAndType(onboardingSelectors.LoginpasswordInput, "password");
+  cy.clearAndType(onboardingSelectors.loginPasswordInput, "password");
   // cy.intercept('GET', '/api/organizations').as('org')
   cy.get(commonSelectors.signUpButton).click();
   cy.wait(2000);
@@ -375,7 +375,7 @@ export const inviteUserWithUserRole = (
 
   cy.wait(1000);
   fetchAndVisitInviteLink(email);
-  cy.clearAndType(onboardingSelectors.LoginpasswordInput, "password");
+  cy.clearAndType(onboardingSelectors.loginPasswordInput, "password");
   cy.get(commonSelectors.signUpButton).click();
   cy.wait(2000);
   cy.get(commonSelectors.acceptInviteButton).click();
